+++ priv/repo/migrations/20260407000000_create_workflow_instances.exs
@@ -2,5 +2,5 @@ defmodule HephaestusEcto.TestRepo.Migrations.CreateWorkflowInstances do
   use Ecto.Migration
 
   def up, do: HephaestusEcto.Migration.up()
-  def down, do: HephaestusEcto.Migration.down()
+  def down, do: HephaestusEcto.Migration.down(version: 1)
 end